+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- `itmdump` tool that parses instrumentation packets from the stimulus port 0 and dumps the payload
-  to `stdout`. 
+- `itmdump` tool that parses instrumentation packets from the stimulus port 0
+  and dumps the payload to `stdout`. 
 
 [Unreleased]: https://github.com/japaric/itm/compare/v0.1.0...HEAD